fix: Remove update checking when inside Docker
This is done by introducing a go build flag.
