+++ update_checker_docker.go
@@ -1,0 +1,7 @@
+//go:build docker
+
+package main
+
+func checkForUpdates() {
+	// do nothing :)
+}
